delegated KCQL properties handling to ProducerProvider
