+++ java-connectors/kafka-connect-azure-eventhubs/src/main/java/io/lenses/java/streamreactor/connect/azure/eventhubs/source/ProducerProvider.java
@@ -1,7 +1,6 @@
 package io.lenses.java.streamreactor.connect.azure.eventhubs.source;
 
 import io.lenses.java.streamreactor.connect.azure.eventhubs.config.AzureEventHubsConfig;
-import io.lenses.java.streamreactor.connect.azure.eventhubs.config.SourceDataType.KeyValueTypes;
 import java.util.concurrent.BlockingQueue;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 
@@ -11,6 +10,5 @@ import org.apache.kafka.clients.consumer.ConsumerRecords;
 public interface ProducerProvider {
 
   BlockingQueueProducer createProducer(AzureEventHubsConfig azureEventHubsConfig,
-      BlockingQueue<ConsumerRecords<Object, Object>> recordBlockingQueue,
-      KeyValueTypes keyValueTypes);
+      BlockingQueue<ConsumerRecords<Object, Object>> recordBlockingQueue);
 }
